Move ROM::Adapter::Memory to ROM::Memory
Aligns the memory repository with how all others are organized.
