--- lib/rom/adapter/memory.rb
@@ -1,1 +1,0 @@
-require 'rom/adapter/memory/repository'
