+++ lib/rom/memory.rb
@@ -1,0 +1,1 @@
+require 'rom/memory/repository'
