+++ spec/integration/mappers/deep_embedded_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-require 'rom/adapter/memory'
+require 'rom/memory'
 
 describe 'Mappers / deeply embedded tuples' do
   let(:setup) { ROM.setup('memory://test') }
