+++ spec/integration/mappers/symbolizing_attributes_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-require 'rom/adapter/memory'
+require 'rom/memory'
 
 describe 'Mappers / Symbolizing atributes' do
   let(:setup) { ROM.setup('memory://test') }
