+++ spec/integration/multi_repo_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-require 'rom/adapter/memory'
+require 'rom/memory'
 
 describe 'Using in-memory adapter for cross-repo access' do
   let(:setup) do
