+++ spec/test/memory_repository_lint_test.rb
@@ -1,4 +1,4 @@
-require 'rom/adapter/memory'
+require 'rom/memory'
 require 'rom/adapter/lint/test'
 
 require 'minitest/autorun'
@@ -7,7 +7,7 @@ class MemoryRepositoryLintTest < Minitest::Test
   include ROM::Adapter::Lint::TestRepository
 
   def setup
-    @repository = ROM::Adapter::Memory::Repository
+    @repository = ROM::Memory::Repository
     @uri = "memory://localhost/test"
   end
 end
@@ -17,6 +17,6 @@ class MemoryAdapterDatasetLintTest < Minitest::Test
 
   def setup
     @data  = [{ name: 'Jane', age: 24 }, { name: 'Joe', age: 25 }]
-    @dataset = ROM::Adapter::Memory::Dataset.new(@data)
+    @dataset = ROM::Memory::Dataset.new(@data)
   end
 end
