+++ spec/unit/rom/memory/storage_spec.rb
@@ -1,14 +1,14 @@
 require 'spec_helper'
-require 'rom/adapter/memory/storage'
+require 'rom/memory/storage'
 
-describe ROM::Adapter::Memory::Storage do
+describe ROM::Memory::Storage do
   describe 'thread safe' do
     let(:threads) { 4 }
     let(:operations) { 5000 }
 
     describe 'data' do
       it 'create datasets properly' do
-        storage = ROM::Adapter::Memory::Storage.new
+        storage = ROM::Memory::Storage.new
 
         threaded_operations do |thread, operation|
           key = "#{thread}:#{operation}"
@@ -21,7 +21,7 @@ describe ROM::Adapter::Memory::Storage do
 
     describe 'dataset' do
       it 'inserts data in proper order' do
-        storage = ROM::Adapter::Memory::Storage.new
+        storage = ROM::Memory::Storage.new
         dataset = storage.create_dataset(:ary)
 
         threaded_operations do
